Remove hardcoded created_by user from returns methods
This applies only to some deprecated blocks and does not
change the new, working logic.

This means that if those methods were used by some store or
extension, they will now don't have the creator set instead
of having a specific user that could even be not present in the
database, causing the creator to be nil anyway.