Change new order creation link
When a user has no orders, the UI displays a link to create a new order.
This link creates a new order, but does not associate the order with the
user that the admin user was redirected from. This change will set the
new order to be for that user by default.