add redirection for offsite payments
`CheckoutController` never had the concept of redirecting the user to the
"offsite" payment method site.
This situation was handled in various bad ways in the past, mostly with the
substitution of the checkout submit button with a link tag, using some JS.

This commit changes `CheckoutController` update method to check if the active
payment wants to redirect to its site, before proceeding to the next checkout
step.