Introduce redirect_url methods
The `Payment` class gets a new method called `#redirect_url`.
In this basic implementation it just delegates to the payment method, but this
interface can be used as a base for different setups. For example, some
payment methods may need to use, or even persist, a dynamically generated
redirect URL.