Set match_path on stock items MenuItem
Without setting match_path the tab helper uses the current controller
name as criteria to activate the tab in the menu. Since the stock items
controller can be accessed for a single product or all products we need
to make sure it only matches the /stock_items path for the Stock menu
tab.

Closes #2950