Move build.sh to bin/build
Using bin/ for internal utils is the default for both Rails and Bundler.
The script has now no extension and is marked as executable improving its
usability (the user doesn't need to know how to deal with each extension) and
maintainability (it's easy to switch to a ruby script just changing the
shebang line).

The doc is updated and env variables are passed with `env` making life easier
for some non-bash shells such as fish-shell.