Allow :state to be a permitted checkout attribute
Using the API can be problematic when updating the order once it has
made it to the confirm step. This is due to calling next when
updating the order when it is in the confirm step. Passing in a new
state fixes this and flows through the state machine starting from
the passed in state.