Remove jquery and rails-ujs requires from backend.js
It's already included in spree/backend/all.js that is generated
by the install command and injected into apps.

Having those same requires in both places is not creating issues since
sprockets is smart enough to understand that and add them once:

https://github.com/rails/sprockets#require

By doing this way we can let users choose which ujs they want to use
since they can just change the value in the spree/backend/all.js that
we are injecting in their application.