Skip setting default state when not included in available countries
Closes #1205. In the case that default country is not included within 
available
countries, this commit leaves the states select empty, where previously 
was
filled in with default country states which didn't match with any 
available
country. The state select is later filled in by Select2, which knows 
which
country option is selected. Server side we can't assume that the first 
country
option is selected, because it doesn't play well with client side 
features, e.g.
autocomplete.