Remove the query counter check
As pointed out by the core team, this spec might be too hazardous in
the future (ex: an update of acts_as_nested_set that cause the code to
make 1 additional query).

I kept the spec to check that the API is correctly handling multiple IDs
as it's an existing feature.