Deprecate solidus:migrations:rename_gateways up and down
This was needed for the v2.2 -> v2.3 upgrade and it should not be
needed anymore.

This also changes the database migration since we are sure it will
never execute any operation after the upgrade. Changing the migration
allows us to deprecate the class responsible for actually doing the
change, which is the same called in the deprecated rake task.

For the original implementation see:

https://github.com/solidusio/solidus/pull/2001/files