Enables a demo using Docker to be ran locally
This commit adds a Dockerfile and entrypoint in
the lib/demo folder, as well as useful documentation
in the README file.

This demo allows users to test out the latest version
of solidus using Docker locally. This setup is very
flexible and provides another option besides using
Heroku or manually installing everything.