Remove destructive actions from migration 20180710170104
Removing tables and renaming columns are destructive actions as they change the
DB schema in a way that is likely to be not compatible with previous code.

This is especially dangerous when deploying multi-instance applications as
migrations can run before all instances code is synced. When this happens,
these instances may experience severe downtime due to high error rates.

So destructive changes should be done in subsequent deploys: first add the new
table and columns, then later remove the stale ones.

Table `spree_store_credit_update_reasons` and its column `update_reason_id`
will need to be removed in a future release of Solidus.