Add migration to drop table/column from `20180710170104`
Migration `20180710170104` has been reworked with PR#3109 in order to avoid
risky situations for multi-instance Solidus installations.

The table `spree_store_credit_update_reasons` and the column `update_reason_id`
from table `spree_store_credit_events` can be either still be around (this
happens when migration `20180710170104` was run after the mod) or not, so we
need to check the presence of both in advance.

Also, since on rollback the older version of `20180710170104` would create this
table then it cannot be created here in a `down` method.

When using the updated version of migration `20180710170104`, the table creation will happen in that migration.