Remove stock locations sorter from main shipping menu in guides
There's no need to keep that item in the menu since it's a
technical implementation detail and could be confusing for someone
who is not familiar with how the sorter works. We are linking
stock locations sorter and filter into the Split Shipment page, which
is the context where those class make sense to be mentioned.