Fix failing spec, add spec to expose `void` behavior
When a single payment exists for the order, and the payment is marked as void,
then the order payment state changes to `Failed`.

When there are multiple pending payments for the order, if only one is marked
as void the order payment state is still `Balance Due`. If all the payments are
marked as void then the order payment state changes to `Failed`.