Deprecate @payment_sources ivar from checkout payments controller
This method returns payment sources of the current user. It is no more
used into our frontend. We used to assign the content of this method
into an ivar (@payment_sources) into the checkout payment step. This
method is here only to be able to deprecate this ivar and will be removed.