Tentative fix for flaky specs
By checking that select2 has correctly been updated before
submitting the form, we ensure that we are passing the right
data to the server.

I also update the portion of the page where we look for the flash
message so it will check for it after the element will be visible
on the page (it is displayed after page load with js).