Ensure return from CSS function
According to the `sassc` compiler, all CSS functions should ensure that
they do not end without a `@return` statement. The `very-light` function
has the possibility that it could exit without calling such a statement,
so we add a catchall `@return` at the end to return the unmodified color
value.

We need this commit since sass-rails is a wrapper for sassc-rails
starting from version 6.0. This means that new apps (or apps that upgrade
that gem) will use sassc by default.

https://github.com/rails/sass-rails/pull/424