Improve fill_in_with_force spec helper
At the moment, after the input is filled with value=, the change
event is not fired.

If we had change listeners on that input they will not be executed.

For example we have a .ccType hidden field that is filled with the
type of the card added and this did remain empty since the backbone
view that was responsible to update it after the card number was
added did not get any trigger.