Set users breadcrumbs from controller to fix bug
This commit makes the `add_breadcrumb` method available in all
controllers to allow to define breadcrumbs from there rather than from
views. It also moves the definition of the breadcrumbs for all the
users' pages from the views to the controller.

Setting the breadcrumbs in the `before_action` of the controller,
ensures objects haven't been changed by the logic of the action yet,
thus avoiding bugs like solidusio/solidus#3149.

This commit includes a regression test for the aforementioned bug, and
tests for the new breadcrumbs rendering logic.