Prevent promotion rule to show user multiple times
The ransack query used to fetch users for the promotion rule dropdown
checked for attributes that may return repeated results. This adds the
`distinct` flag in the result's call to prevent repeated records to be
shown.

Although this issue was seen in the promotion rule's dropdown, the bug
could've been happening in different places since it lives in the resource
controller.