Validate store credit amount is not used before discarding
We already had this but with the last version of discard the validation
is no more performing.

I think it's something related to this commit
https://github.com/jhawthorn/discard/commit/dfdd984b37bf3d646b276d68e47f7f998c6aaad9
which is included in the last release.

Before the release, calling discarded? on the line:

validate :validate_no_amount_used, if: :discarded?

was returning true, even if the record change wasn't yet persisted. We
don't need that line anymore since now records can't be discarded
without passing the after_discard validation so we should not have any of
those. Anyway, even if that happens we don't want to block other updates
on those records.