Fix the way to call the deprecated variables mixin
In the relative previous commit, the old numbered color variables has
been re-added for print a custom warning message; actually the message
was printed every time and in any case, and it was not so helpful. Now,
the warning msg is printed only if a variable is declared in
`_variable_override.scss`.