Fix specs for `reset_spree_preferences`
`Spree::Config.mails_from` is set in `dummy_app.rb`, but when using the method
`reset_spree_preferences` that change is lost.

In order to successfully test the method by veryfing that preference, it's
necessary to execute `reset_spree_preferences` before checking the value for
`mails_from`, so the before/after values are the same.