Add instructions on how to extend ransackable attributes
When we add a new input to the search form in the admin, we also
need to add the new attribute to the whitelisted attribute that ransack
can use. This is done for security reasons.

Having this placeholder can help users figuring out what they need
to do when the the search is not working.