Upload new sample images
Refresh sample images of the store demo, with items branded by Solidus and by a
redesigned version of the Ruby logo.

Update the image sizes to support the retina screens (is not an appropriate
way, maybe we need an helper that load bigger images only on retina screens),
except the smaller one preview that is in use on the product page when there
are many variants, and in admin area on product list; the issue is that on the
product page there is any max-width defined for the thumbnail list items.