Do not disable belongs_to_required_by_default app wide
With this ActiveRecord class setting we make the Rails 5.1 default
setting obsolote and turn this off app wide.

   Rails.application.config.active_record.belongs_to_required_by_default
   # => true
   Spree::Base.belongs_to_required_by_default # => false
   ActiveRecord::Base.belongs_to_required_by_default # => false

The fun fact is, the tests all pass with this removed.
But this has the potential to break stores if their data is
not consistent.

We have several options to solve this

1. We tell our users to clean up the database before running the next
Solidus update where this included.
2. We include this in a mayor version update
3. We explicitely make all our `belongs_to` associations `optional`

I am in favor of the latter