Bump CircleCi gems cache key
The docker image that we are using for mysql has been recently
updated.

https://hub.docker.com/r/circleci/mysql

This update changed some internal path or library that were cached
into our gem cache in circleci for `mysql2 0.5.2`. This happens
because native extensions are cached along with the gems and are not
rebuilt at every job run.

To clear CircleCI cache we can only change the cache key to rebuild
it from scratch, as described in their docs:

https://circleci.com/docs/2.0/caching/#clearing-cache