Only display Store Credits tab with permission
This change hides the "Store Credits" tab if the user does not
have permission to display store credits for the customer.

Within larger applications it's desirable to scope user permissions
by the store. The admin users are only allowed to manage a customer's
store credits if they're a manager of the customer's store.

So the permissions check needs to be done on a new users store credit
rather than the class itself for a proper check.

https://github.com/CanCanCommunity/cancancan/wiki/Nested-Resources#accessing-parent-in-ability