Click somewhere else on the page before submit
Without this line we have a flaky spec probably due to select2 not closing
its fixed overlay correctly.

We've notice that the submit button wasn't pressed at all when we had
failures. This means something is preventing the click to happen.
The select2 plugin has a lot of code that interacts with other page
elements preventing clicks so that could be the root cause.

Clicking anywhere in the page before submit apparently solves the issue.