Add docs for partials that need to be provided
Due to !2513 we need to provide a partial for solidus_api to render a
payment method. This is much like how payment methods need to provide
partials for the solidus_backend and solidus_frontend. The documentation
does not yet reflect this. Include documentation for providing this
partial on the api side.