Allow splitting shipments when not tracking inventory
When Spree::Config.track_inventory_levels is false we are
not supposed to change inventory levels when shipping orders.

At the moment though, if we try to split a shipment with the
following scenario:

- track_inventory_levels is false
- there is no inventory in the second stock location for the
  related stock item

we end up creating a second shipment with items backorderable,
making the second shipment impossible to ship.

This is because the FulfilmentChanger class at the moment does
not take into account this scenario. But it was currently doing
a lot of things that are not required when inventory leves are
not tracked.

This fix proposes two different paths, driven by a new argument
required in the initialization of the class, that perform a
complex stock movement when we are tracking inventory and a more
simple otherwise.

Not passing this new argument to the class' `new` method is
deprecated, assuming the old behavior for backward compatibility.