Add specs for Spree::Calculator::Shipping::PriceSack
The previous specs file was actually testing the main
Spree::Calculator::PriceSack instead.

The allow_any_instance hack is required since while building
the package content, no matter what is the price of the variant,
the line item amount is always 10. This is due to how the factory
is defined:

https://github.com/solidusio/solidus/blob/b76c1dd67642245778c019b15b3a10ef22b1f50c/core/lib/spree/testing_support/factories/line_item_factory.rb#L9