Remove useless Spree::ShippingCalculator interface method
We don't need to check and raise if that method is not defined
since it's not used in our codebase for that kind of calculator.

This is the reason why also specs are useless.

We are able to remove this method without a deprecation warning
because it's an abstract method and, if users are overriding
it into their own calculators that inherit from this one their
application will be perfectly safe.