Pass stock location to inventory unit factory
When creating fulfilled stock packages in specs, this passes the given
transient stock location to the inventory units that get built. This
prevents the inventory unit factory from building unique stock locations
for each inventory unit that gets built, and ensures that the stock
package is using the same stock location as the inventory units it
contains.