Reload product to fix option value lookup
After adding the variant & option value includes within commit
81ddee979595150df708299e484f8e6ad583ec9c the failure to find option
value bug moved to the solidus_girly product. I have no idea why, but
simply reloading the product fixes the issue which was my original
solution before adding the includes. This doesn't fix whatever the
underlying issue is, but is a simple hack that seems to resolve the
flakey failures during the sample loading.