Use RESTful routing for users' API key management
This commit refactors how we are using routes to use RESTful
actions, moving the non-RESTful endpoint into a separate
controller.

This has the benefit to have a more organized codebase, but
also to improve the granularity of how we can set permissions.
For example, now we can have a permission set that does not
allow to change or revoke users' API key but allows management
of other users things. For this reason, a new ability is required
and has been added in the provided UserManagement permission set.

This commit also deprecates the old controller actions, printing
a message in the log of users that are still using the old actions.

The controller tests for the old actions have been deprecated as
well, and copied over the new controller specs file but it is
only checking controller related stuff now, since the real
functionality is already covered by unit and feature specs
elsewhere.