Explicitly set Money rounding to ROUND_HALF_EVEN
This is the current default value, but it will change in
the Money gem in some future release (v7).

People can start adhering to that value by setting

Money.rounding_mode = BigDecimal::ROUND_HALF_UP

in their own application. For now, this is what we used so far
and we should not change it for existing applications.

Ref: https://github.com/RubyMoney/money/pull/883