Set raise_with_invalid_currency to false for new app
This ensures that the DummyApp that we use for specs and new
applications will use the new default behavior, while preserving
the old behavior for existing applications that are just upgrading.

Setting this new default for the DummyApp allows us to restore a
pending spec that was skipped for the inconsistent raise behavior
that we want to deprecate.