Standardize spree event context parameter
When calling `Spree::Event.fire 'event'`. It is possible to pass in
parameters. The parameter is usually the context to the event such as an
order in order_finalized or reimbursement when a reimbursement has been
created.

In order to standardize this context rather than calling
`Spree::Event.fire 'order_finalized', order: @order`, call
`Spree::Event.fire 'order_finalized', context: @order`.

This will also allow subscribers to always know which key will contain
the context to the event.