Let the order transition to Returned after return items are all Received
Apparently the state_machines library can halt callback chains
without telling anything to anyone. That was happening in some
cases when ReturnItem#attempt_accept was failing. In those cases, just
to increase confusion, the acceptance state-machine is interacting
with the reception state-machine.

Additionally, the introduction of the accessor for
skip_customer_return_processing was fixing one code path (the one
still visible in the original PR) but was disrupting another one,
namely marking the order as "returned" after the return-item
state-machine was delivered the "receive" event.