Create a valid reimbursement in sample
The current code is not creating a valid reimbursement since it's
created when the associated order is still not in the riht state
(paid and shipped).

The amount of the reimbursement was also wrong, since the return
item created had an exchange variant associated, making the total
of the calculated reimboursement 0.

This was mainly used to allow seeing the reimbursement email in the
preview (with #1146), and it was working fine for that purpose but it
is creating an inconsistent order in the backend running samples.

Retruns are already complex and messed up, we need to avoid showing
wrong examples of code and order states to users.