Reimburse sample order after its creation
The sample order's reimbursement is created correctly but
`perform!` is never called on it. This means that the instance
of Spree::Reimbursement is persisted, but no action is taken
on the reimbursment for this return (no payment refund is made).

This commit change how we create the reimbursement, the provided
method #build_from_customer_return already accept return items,
create the reimbursement taking what is needed by the customer
return instance and also perform the acutal refund.

To perform a reimbursement we need to pass the user that is taking
this action. When solidus_auth_devise is there we can assume we
have an admin user, at least this is what happens in the sandbox
envirnoment. When it's not used we can just try to fetch/create a
dummy user.