Scope image by product_id or variant_id in ImagesController
The image route is under the product and variant resources:
`/api/products/:product_id/images/:id`
`/api/variants/:variant_id/images/:id`

When reading or perfoming any change on the product/variant image, the controller
should scope the product or the variant when finding the image id.
The risk in only find by image id is to delete or read an image which belongs to
another product.
The route path give an expectation that the controller must meet:
a request to a product or variant image.