Allow option type to be nil when option value delegates option_type_name to it
Spree::OptionValue#name is delegated to Spree::OptionType with option_type
prefix (option_type_name).

https://github.com/solidusio/solidus/blob/v2.9.5/core/app/models/spree/option_value.rb#L17

The json response of Spree::Api::OptionValuesController uses the
*option_value_attributes helper variable which has :option_type_name.
If for some reason an option value is created without option_type the above
controller will return a 500 error because the name delegation.

This commit also remove check_option_values from the option_values_controller_spec.rb
because is not used anywhere in the file.