Replace `Spree::Event#name_with_suffix` with `adapter#normalize_name`
`#name_with_suffix` is not generic enough to stay in the `Spree::Event` class,
as different adapters may not have the concept of a name suffix, while, on the
other hand, may need some different naming manipulation.

So, event name manipulation is now delegated to the specific eventadapter. As
we currently have only `ActiveSupportNotifications`, the existing code from
`Spree::Event` has been moved in that adapter class.