Deprecate usage of spree:install generator
This allows the old generator to continue working and
it will print a deprecation warning at the end of the
process to inform users that they need to switch to the
new generator.

I think that currently this is only used into other extensions
or support gems, like solidus_dev_support.