Rescue from invalid state machine exception in base controller
The rescue is moved on Spree::Api::BaseController so that it can be inherited
from those controllers that deal with state machine transition.
This commit also adds a test to ensure OrdersController#cancel does not return
500 when trying to transition to `cancel` an order that is not completed.

When an order is not completed cannot be transitioned.

When called in OrdersController#cancel through `canceled_by` the state machine
exception result in a 500 error.