Make Order state machine configurable
In 7c4f82c21, all relevant state machines except for the order one have
been extracted into modules. This performs the same exercise for the
order state machine. A lot of this code is deeply interlinked with the
rest of the system, so you'd have to take lots of care when replacing
this; but it should be possible nonetheless.

The new state machine is named Spree::Core::StateMachines::Order and
includes the same code of the old one.

The newly named state machine is active on new stores generated from
scratch and in Solidus tests, but old stores will still use the old one,
unless the preference `Spree::Config.use_legacy_order_state_machine` is
explicitly set to `false`.

When using the old named state machine a deprecation warning will be
printed at boot time, as Solidus 3.0 will start using the new order state
machine by default.