Make Payment#gateway_order_id public
This method, although in Solidus codebase is currently called only by the
class that defines it, can be useful in other scenarios (for example when
one wants to use it outside the standard Solidus payment creation flow)
so it should be public.

Also, it represents information that is not consumed only inside the
Spree::Payment model, but ends up stored externally on the payment
provider backend.