Prevent OpenURI redirection and private hosts during API image upload
Paperclip’s CVE-2017–0889 Server Side Request Forgery (SSRF)
vulnerability (https://medium.com/in-the-weeds/all-about-paperclips-cve-2017-0889-server-side-request-forgery-ssrf-vulnerability-8cb2b1c96fe8) was likely introduced in #3573.  This patch
ensures that redirection does not take place in the context of
uploading an image via the api, and ensures that a private address
is not used for the attachment parameter.