First attempt to make images shared with multiple variants
THe idea was to create a two new tables to allow images to belong
to multiple variants, the first one is to be the polymorphic association
and then the M-N table following rails conventions.
It is going to be hard to maintain old functionality due the change of
relationships and variable naming, it basically works without rewriting
the frontend, gallery helped but the trick will be do it backward compatible
with old images when migrating to newer version of solidus, maybe a rake task?