Frontend: Try to add a coupon code only when updating checkouts
The `apply_coupon_code` before filter should only run on update. We
shouldn't allow this expensive code to run on a GET request.

This is probably also a mitigation for an imaginable denial of service
attack, where an attacker could hammer a site with lots of coupon
codes, bringing the site down because of all the work necessary to
check for the code's validity.

This action is already deprecated and removing this functionality
should not create many backward compatibilities issues. Also,
for the reasoning above we can consider this a bug.

Co-authored-by: mamhoff <mamhoff@users.noreply.github.com>