Update order cancellation, sets inventory units to canceled
Before this commit when an order was canceled the inventory units
of the order will remain untouched. So there were canceled orders with
inventory units with the state on_hand or backordered.

This situation was not only inconsistent but can lead to subtle bugs
with the stock management system.

With this commit the inventory units related to an order are
canceled when the order is canceled.

When an order is resume we recalculate the state of the connected
inventory units, based on the actual stocks available.