Deprecate Spree::Event.subscribers and .require_subscriber_files
`Spree::Event.subscribers` is deprecated and its functionality is
replaced by `.subscriber_registry`.

`.subscriber_registry` is not a 100% compatible replacement for
`.subscribers`, as the internal mapping system for event subscribers
is changing completely, but for existing Solidus stores the change
should be transparent enough.

`Spree::Event.require_subscriber_files` is deprecated and delegated
to the subscriber registry object. The method with the same name
defined in `Spree::Event::SubscriberRegistry` is now private, as
there is no use in keeping it public.