Remove deprecated coupon code functionality
It now has a separate endpoint and it's no more
possible to apply a coupon code passing a parameter
when updating the order.

Similarly, we are removing a non-restful action in the
api orders controller in favor of a dedicated controller
that takes care of handling the coupon code functionality.