Add customer orders list page to solidus_frontend
The code is basically a cut & paste of the relevant parts from the gem
`solidus_auth_devise` with minor improvements.

The rationale for this change is that the orders history better belongs
inside the solidus main project, because the feature does not have much
to do with the actual authentication process, and because it allows
projects that have an existing authentication system (ie. don't use
solidus_auth_devise) to have the orders list page feature out of the
box.

In `solidus_auth_devise` the list is on the `/account` page. This page
also includes some minimal account information such as the user's email
and a link to edit the account. This part as well is now in Solidus.