Move "thinking cat" fixture to lib folder
With the exclusion of the spec folder from the packaged solidus_core gem
in 80bea87, all factories that use an image can now not access that
image when loaded from the lib folder of a packaged gem.

The reason for that is that the image referenced in those factories
lives under spec/fixtures - in the spec folder.

This commit does the simplest thing that could possibly work: Move the
image to the lib/fixtures folder instead and adapt the factories so that
they work even if loaded from the packaged gem.