Remove Spree::Confing.use_legacy_order_state_machine
and Spree::Order::Checkout module. We are now using a new
state machine and the approach to configure it is changed.

If you changed that module with monkey patches, the best
approach to be compliant is copy/pasting the provided
state machine into a new module in your store, change its
code to reflect the things you changed with the monkey patch
and use it by setting:

	Spree::Config.state_machines.order = 'Your::Module'

See Solidus Guides for more information.