Avoid saving a refund when process! fails
If we save unsuccessful refund on payment gateway errors
we'll end up having a refund associated to the payment without
the actual refund processed.

Since at the moment refunds have no states to indicate whether
is successful or not, the system is considering it as done and
doesn't allow to perform a new one since the credit allowed amount
for the payment becomes 0:

payment.credit_allowed = payment total - refund amount

and this formula is true also when the refund has not been succesful.

This comes from our past behavior of creating and calling perform! on
refunds at the same time.

By just validating the refund we have the same result. The
"perform_after_create: false" attribute setting has been
moved into the perform! itself. When called on a record still
unpersisted, it will set its value to false to avoid deprecation
warnings and well.. to call perform! again!

A more clean solution would be considering in that formula only
refunds that are completed but at the moment there's no way to
do that. This way we could also keep track of the failed attempts
to perform refunds, saving all the information we get back from the
gateway.