Install solidus without javascript compilers (webpacker)
As taken from edge guides: solidusio/edgeguides#12

Using --skip-webpack-install is only preventing the Rails install
generator from running webacker:install, but the webpacker gem is
still added to the Gemfile causing an error when starting the server:
'Webpacker configuration file not found'

In the context of a fresh app that uses the bare minimum set of tools to
run Solidus we don't need that. People can still add it later if needed.

Co-authored-by: Alberto Vena <kennyadsl@gmail.com>