Fix permissions to see admin menu items
Right now a user without roles like logged users or guests can
see the admin menu partially.

This is because it actually has permission to show zones and
shipping methods and the current code is checking against the
:show ability, so it seems to be legit.

We use :admin for the rest of the menu items checks though, and
this is also what we use in the controller to determine if
we can access that page, see: https://github.com/solidusio/solidus/blob/3c8ffcc34f9248b286a9d4ca94d1f9a3197ac7b2/backend/app/controllers/spree/admin/base_controller.rb#L29

Test setup lines have been added to be sure the links are populated
correctly in the menu.