Remove deprecated cancan abilities
With 915ffa3149fd1a532e22c55bb9fd715292991b6b we correctly switched
a custom cancan alias to use the correct one and print a deprecation
warning when the wrong one was used.

But we forgot to update the code to reflect this change. No deprecation
warnings were raising because we have the use_custom_cancancan_actions
set to false on new apps.

I switched that preference to true locally and specs started to raise
deprecations warnings, signaling where are using the wrong alias and
this commit is the result of that investigation.