Update taxon icon to use present instead of exists
`Exists` on a paperclip object attempts to access the image where
it's stored, whereas present just checks if the object exists on
Solidus.

Using exists here causes an error if the image doesn't exist. Luckily,
paperclip double checks that the image filename is present before
attempting to fetch it from the storage service, but paperclip_cloudinary
does not do this, meaning that any store using paperclip_cloudinary is
unable to view the admin taxon edit page for taxons without icons.