Add Address name data migration rake task
After solidusio#3908 the table spree_addresses includes the name
field, so it can be populated with historical data already present
in the table by joining the content of firstname and lastname.
Running `rails solidus:migrations:migrate_address_names:up` will
take care of the job.

The task is arranged in steps that require user input in order to
proceed:

  * the task will exit no-op when no record with blank name is
    found;
  * if records are found, the user is asked to confirm before
    proceeding;
  * the user is now asked for the record batch size to be used.
    My tests showed that 100'000 records per batch should balance
    the necessity to avoid locking records for long and have the
    script complete without too much overhead;
  * before migrating data, the user is asked to confirm once again.

Running this task is required when upgrading to Solidus 3.0.