Use Spree::Order.user email format validators on Spree::EmailValidator
This patch solves a problem where the e-mail for Spree::Order is copied from Spree.user_class, specifically when spree_auth_devise is used.
Without it, Spree::Order might be invalid after persistence when the Spree.user_class email does not match the standard EMAIL_REGEXP of Spree::EmailValidator

ActiveRecord's reflection is used to reflect upon the record's user association. The advantage is that this works whether spree_auth_devise or another user gem is used.