Add a post install/update message
We already introduced this post-install message in 2.11.5
when we started requesting users to run a rake task after
bumping Solidus.

Even if we don't have any specific rake task here in 3.0 we
can keep this convention so we always have a place where to
put this kind of tasks.

Also, we can be consistent between versions in terms of how
we request people to upgrade Solidus:

1. Change Gemfile
2. Run `bin/rails solidus:upgrade`

Even if it does nothing special, I think it's a positive thing
to have.