Fix order checkout flow when confirm step is missing
The Spree::Order model has a configurable checkout flow where steps can
be customised. One of the configurable steps is 'confirm'. As a result,
the 'confirm' step is optional.

The code defining the transition for the 'complete' event was still only
transitioning from the 'confirm' state, rather than from the last
configured checkout flow state. This was causing a problem in case the
'confirm' step was actually missing.

This commit removes the hardwired 'confirm' state reference and obtains
the last configured flow state instead.